dang lam form chongoi

+++ src/main/java/com/demo/DAO/PhongDao.java
@@ -96,7 +96,7 @@ public class PhongDao {
             pst.setString(3, gioBatDau);
             rs = pst.executeQuery();
             while (rs.next()) {
-                list.add(new ChiTietGhe(rs.getInt(1),rs.getString(2), rs.getString(3), rs.getInt(4), rs.getDouble(5), rs.getInt(6)));
+                list.add(new ChiTietGhe(rs.getInt(1),rs.getString(2), rs.getString(3), rs.getInt(4), rs.getFloat(5), rs.getInt(6)));
             }
         } catch (SQLException ex) {
             Logger.getLogger(PhongDao.class.getName()).log(Level.SEVERE, null, ex);
